added the add to cart function to the list-view
