+++ js/index.js
@@ -1,5 +1,5 @@
 import { fetchJackets, createTitle1, createTitle2, showLoadingIndicator } from "./global.js";
-import { getProductsFromCart } from "./cartfunction.js";
+import { eventSaveLocally } from "./cartfunction.js";
 
 export const sliderSection = document.querySelector(".product-slider");
 const adSection = document.querySelector(".ad-section");
@@ -64,45 +64,11 @@ async function displayProducts() {
   </div>`;
   adSection.appendChild(adContainer);
 
-  //move to its own js-file/or global after completing
+  //ADD ITEMS TO CART
   const addToCartButton = document.querySelectorAll(".shopping-bag");
 
   addToCartButton.forEach((cartButtons) => {
     cartButtons.addEventListener("click", eventSaveLocally);
   });
-
-  function eventSaveLocally() {
-    event.target.classList.toggle("shopping-bag_icon-empty");
-    event.target.classList.toggle("shopping-bag_icon-added-product");
-
-    const id = event.target.dataset.id;
-    const title1 = event.target.dataset.title1;
-    const title2 = event.target.dataset.title2;
-    const img = event.target.dataset.img;
-    const price = event.target.dataset.price;
-    const description = event.target.dataset.description;
-    const sizes = event.target.dataset.sizes;
-
-    const sizeArray = sizes.split(",");
-
-    const productsCurrentlyInCart = getProductsFromCart();
-
-    const productExists = productsCurrentlyInCart.find(function (item) {
-      return item.id === id;
-    });
-
-    if (!productExists) {
-      const item = { id: id, title1: title1, title2: title2, img: img, price: price, description: description, sizes: sizeArray };
-      productsCurrentlyInCart.push(item);
-      addToCart(productsCurrentlyInCart);
-    } else {
-      const newProducts = productsCurrentlyInCart.filter((item) => item.id !== id);
-      addToCart(newProducts);
-    }
-  }
 }
 displayProducts();
-
-function addToCart(product) {
-  localStorage.setItem("inCart", JSON.stringify(product));
-}
