+++ js/list.js
@@ -1,4 +1,5 @@
 import { fetchJackets, createTitle1, createTitle2, showLoadingIndicator } from "./global.js";
+import { eventSaveLocally } from "./cartfunction.js";
 
 const listSection = document.querySelector(".product-list");
 
@@ -16,9 +17,8 @@ async function displayProducts() {
     productContainer.classList.add("product");
 
     productContainer.innerHTML += `
-    <a href="checkout.html" class="shopping-bag">
-      <img src="images/icons/shopping-bag.png" alt="link to shopping-bag" />
-    </a>
+    <div class="shopping-bag shopping-bag_icon-empty" alt="link to shopping-bag" data-img="${product[i].image}" data-id="${product[i].id}" data-title1="${title1}" data-title2="${title2}" data-description="${product[i].description}" data-price="${product[i].price}" data-sizes="${product[i].sizes}">
+    </div>
     <a href="product.html?key=${product[i].id}" class="product-link">
       <img src="${product[i].image}" alt="${product[i].description}" class="jacket" />
     </a>
@@ -27,15 +27,17 @@ async function displayProducts() {
       <a href="product.html?key=${product[i].id}" class="black">${title2}</a>
       <p class="productprice">$${product[i].price}</p>
     </div>`;
-    productContainer.addEventListener("click", () => {
-      window.location.href = `product.html?key=${product[i].id}`;
-    });
-
     listSection.appendChild(productContainer);
-    
-    if (i === 9) {
+
+    if (i === 8) {
       break;
     }
   }
+  //ADD ITEMS TO CART
+  const addToCartButton = document.querySelectorAll(".shopping-bag");
+
+  addToCartButton.forEach((cartButtons) => {
+    cartButtons.addEventListener("click", eventSaveLocally);
+  });
 }
 displayProducts();
